fixing mistakes made during a three day meth binge
Meth is a hellovadrug man.  Lemme tell you.  That shit will keep you up for fucking days cleaning the same damn spot on the carpet just for you to realize that it's not the carpet anymore but that you had scrubbed your way all the way down to the sub floor.  I'd only recommend using that shit under close adult supervision.